Implement setting per-channel volumes
Backport parts of https://invent.kde.org/plasma/plasma-pa/-/commit/12c07e355b8cebaea61413bd870e27904b3eada6

+++ src/volumeobject.cpp
@@ -57,9 +57,9 @@ QStringList VolumeObject::rawChannels() const
     return d->m_rawChannels;
 }
 
-QVector<qreal> VolumeObject::channelVolumes() const
+QVector<qint64> VolumeObject::channelVolumes() const
 {
-    QVector<qreal> ret;
+    QVector<qint64> ret;
     ret.reserve(d->m_volume.channels);
     for (int i = 0; i < d->m_volume.channels; ++i) {
         ret << d->m_volume.values[i];
